requirements.txt
appdirs==1.4.4
attrs==21.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
binwalk==2.3.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
fonttools==4.29.1
fs==2.4.12
gast==0.5.2
html5lib==1.1
idna==3.3
image-similarity-measures==0.3.6
imageio==2.28.1
imutils==0.5.4
kiwisolver==1.3.2
lazy_loader==0.2
lxml==4.8.0
lz4==3.1.3+dfsg
matplotlib==3.5.1
mpmath==0.0.0
networkx==3.1
numpy==1.24.3
olefile==0.46
opencv-python==4.7.0.72
packaging==21.3
phasepack==1.5
Pillow==9.0.1
ply==3.11
pycron==3.0.0
pycryptodome==3.17
pyFFTW==0.13.1
Pygments==2.11.2
PyGObject==3.42.1
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyqtgraph==0.12.4
python-dateutil==2.8.1
python-magic==0.4.24
pythran==0.10.0
pytz==2022.1
PyWavelets==1.4.1
PyYAML==5.4.1
requests==2.25.1
scikit-image==0.20.0
scipy==1.8.0
six==1.16.0
soupsieve==2.3.1
sympy==1.9
tifffile==2023.4.12
ufoLib2==0.13.1
unicodedata2==14.0.0
Unidecode==1.3.3
urllib3==1.26.5
webencodings==0.5.1
